Persistent Runner Worker (#41)
This pull request implements a persistent runner, which maintains a per
device persistent subprocess that caches compiled solutions (to avoid
re-compilation for every workload during benchmarking) and exposes a
fast_benchmarking option inside the benchmark class, users can opt for a
faster benchmarking process with less isolation, or use the slower
multi-process runner for safe benchmarking.

NOTE: we recommend using the multi-process runner on smaller datasets or
unproven kernel solutions

The persistent workers will conduct health checks to ensure healthy gpu
context before running testing.

In addition, we deem a solution that has raised a dtype error, compile
error, or runtime error for over 3 consecutive trials as a broken
solution, and skip it for the rest of the benchmarking process.